+++ R/extract.R
@@ -22,10 +22,10 @@ extract <- function(models,
                     gof_map = modelsummary::gof_map,
                     gof_omit = NULL,
                     add_rows = NULL,
-                    add_rows_location = NULL,
                     stars = FALSE,
                     fmt = '%.3f',
                     estimate = 'estimate',
+                    add_rows_location = NULL,
                     ...) {
 
     # models must be a list of models
@@ -40,7 +40,8 @@ extract <- function(models,
     sanity_coef_omit(coef_omit)
     sanity_gof_map(gof_map)
     sanity_gof_omit(gof_omit)
-    sanity_add_rows(add_rows, add_rows_location, models)
+    sanity_add_rows_location(add_rows_location) # deprecated 
+    sanity_add_rows(add_rows, models)
     sanity_stars(stars)
     sanity_fmt(fmt)
     sanity_estimate(estimate)
@@ -132,33 +133,30 @@ extract <- function(models,
     gof <- gof[order(idx, gof$term),] 
 
     # add_rows: this needs to be done after sorting and combining to preserve
-    # user-selected row order
+    # user-selected row order. We need the 'section' information otherwise we
+    # won't know where to draw the line which separates estimates from gof. 
     if (!is.null(add_rows)) {
 
-        if (is.null(attr(add_rows, 'section'))) {
-            loc <- 'bottom'
-        } else if (attr(add_rows, 'section') == 'bottom') {
-            loc <- 'bottom'
-        } else if (attr(add_rows, 'section') == 'middle') {
-            loc <- 'middle'
+        if (inherits(add_rows, 'list')) {
+            add_rows <- do.call('rbind', add_rows) %>%
+                        data.frame(stringsAsFactors = FALSE) %>%
+                        setNames(colnames(gof)[2:ncol(gof)]) %>%
+                        dplyr::mutate(section = 'bottom', position = 1e6)
         }
 
-        rows <- do.call('rbind', add_rows) %>%
-                data.frame(stringsAsFactors = FALSE) %>%
-                setNames(colnames(gof)[2:ncol(gof)])
- 
-        if (loc == 'middle') {
-            rows <- rows %>% 
-                    dplyr::mutate(group = 'estimate', statistic = 'estimate')
-            rows <- rows[, colnames(est)]
-            est <- est %>% 
-                   tibble::add_row(rows, .before = add_rows_location)
-        } else if (loc == 'bottom') {
-            rows <- rows %>% 
-                    dplyr::mutate(group = 'gof', statistic = '')
-            rows <- rows[, colnames(gof)]
-            gof <- gof %>% 
-                   tibble::add_row(rows, .before = add_rows_location)
+        for (i in 1:nrow(add_rows)) {
+            newrow <- add_rows[i, ] %>% dplyr::select(-section, -position)
+            if (add_rows$section[i] == 'bottom') {
+                newrow <- newrow %>% dplyr::mutate(group = 'gof') 
+                gof <- gof %>%
+                       tibble::add_row(newrow, .before = add_rows$position[i])
+            } else {
+                newrow <- newrow %>% 
+                          dplyr::mutate(group = 'estimates',
+                                 statistic = 'estimate') 
+                est <- est %>%
+                       tibble::add_row(newrow, .before = add_rows$position[i])
+            }
         }
 
     }