+++ R/modelsummary.R
@@ -49,11 +49,15 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' `omit`. See `modelsummary::gof_map`
 #' @param gof_omit string regular expression. Omits all matching gof statistics from
 #' the table (using `stringr::str_detect`).
-#' @param add_rows list of character vectors, each of length equal to the number
-#' of models + 1.
-#' @param add_rows_location integer or NULL. custom rows will be added to the
-#' bottom of the table if this parameter is NULL, or after the position set by
-#' this integer.
+#' @param add_rows a data.frame (or tibble) with the following columns: 
+#' \itemize{
+#'   \item section (character): insert in "middle" or "bottom" section of the table
+#'   \item position (integer): row position in the section
+#'   \item term (character): string to display under coefficient names
+#'   \item one column per model with the same name as that model with the
+#'         values to insert (some models can be omitted). 
+#'   \item See the examples section of this documentation and an example.
+#' }
 #' @param title string
 #' @param notes list or vector of notes to append to the bottom of the table.
 #' @param estimate character name of the estimate to display. Must be a column
@@ -61,13 +65,16 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #' cases, the default value of this argument should not be changed.
 #' @param filename This argument was deprecated in favor of the `output` argument.
 #' @param subtitle This argument is deprecated. Use `title` or the `tab_header`
-#' function from the `gt` package.
+#' @param add_rows_location This argument is deprecated. Use a data.frame as
+#' described in the documentation for the `add_rows` argument.
 #' @param ... all other arguments are passed to the `tidy` method used to
 #' extract estimates from the model. For example, this allows users to set
 #' `exponentiate=TRUE` to exponentiate logistic regression coefficients.
 #' @return a 'gt' table object.
 #' @examples
 #' \donttest{
+#' library(modelsummary)
+#'
 #' # load data and estimate models
 #' data(trees)
 #' models <- list()
@@ -90,6 +97,13 @@ globalVariables(c('.', 'term', 'group', 'estimate', 'conf.high', 'conf.low', 'va
 #'
 #' # title with italicized text
 #' msummary(models, title = gt::md('This is *the* title'))
+#' 
+#' # add_rows: we use `tribble` from the `tibble` package to build a data.frame
+#' # more easily
+#' rows <- tibble::tribble(~term, ~section, ~position, ~Bivariate, ~Multivariate,
+#'                         'Empty row', 'middle', 3, '-', '-',
+#'                         'Another empty row', 'bottom', 1, '?', '?')
+#' msummary(models, add_rows = rows)
 #'
 #' # notes at the bottom of the table (here, the second note includes markdown bold characters)
 #' msummary(models, notes = list('A first note', gt::md('A **bold** note')))
@@ -118,12 +132,12 @@ modelsummary <- function(models,
                          gof_map = modelsummary::gof_map,
                          gof_omit = NULL,
                          add_rows = NULL,
-                         add_rows_location = NULL,
                          title = NULL,
                          notes = NULL,
                          estimate = 'estimate',
                          filename = NULL,
                          subtitle = NULL,
+                         add_rows_location = NULL,
                          ...) {